real-pmap: reliable early abort
real-pmap cancelled futures if it encountered an exception from any
thread, but this depended on a linear traversal of the mapper results.
If, say, the first mapper was in a barrier waiting for the second mapper
to decrement that barrier, and the second thread crashed, then the first
mapper would never return, and real-pmap would stall forever.

This new approach interrupts every thread as soon as any encounter an
error, which prevents these deadlocks, ensures that all threads have
terminated, then rethrows the original exception.