loopr: significant speedup for multi-acc reduce
With multiple accumulators, loopr's :reduce tactic now generates a
reduction which pulls the extra accumulators out into volatiles, rather
than destructuring and re-structuring them each step through the reduce.
This improves performance with 2 accumulators by roughly 40% vs (reduce
(fn [[acc1 acc2] x] ... [acc1' acc2']) ... xs). It's still not as fast
as the iterator tactic, though, which is ~66% faster, so we leave that
as the default.